+++ source/common/modules/markdown-editor/editor-extension-sets.ts
@@ -75,6 +75,7 @@ import { vimPlugin } from './plugins/vim-mode'
 import { projectInfoField } from './plugins/project-info-field'
 import { headingGutter } from './renderers/render-headings'
 import { codeTheme } from './renderers/render-code'
+import { citationTooltips } from './tooltips/citations'
 
 /**
  * This interface describes the required properties which the extension sets
@@ -345,6 +346,7 @@ export function getMarkdownExtensions (options: CoreExtensionOptions): Extension
     footnoteGutter, // Should be after markdownFolding
     urlHover,
     filePreview,
+    citationTooltips,
     backgroundLayers, // Add a background behind inline code and code blocks
     defaultContextMenu, // A default context menu
     softwrapVisualIndent, // Always indent visually